requirements.txt
APScheduler==3.11.2
bcrypt==5.0.0
blinker==1.9.0
boto3==1.42.30
botocore==1.42.30
click==8.3.1
Flask==3.1.2
Flask-APScheduler==1.13.1
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==5.21.0
greenlet==3.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
s3transfer==0.16.0
six==1.17.0
SQLAlchemy==2.0.45
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
Werkzeug==3.1.5

# Testing dependencies
pytest>=7.4.0
pytest-flask>=1.2.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
